backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Marcos Mantovani <marcospiderplays@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
